refactor: 학부 안내, 교과 과정, 전공 이수 표준 형태
