+++ .github/workflows/deploy.yaml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - main
-      - refactor/research
+      - refactor/academics
 
 jobs:
   build:
