+++ app/[locale]/academics/helper/CourseCard.tsx
