+++ app/[locale]/academics/helper/CourseCards.tsx
