+++ app/[locale]/academics/helper/CourseDetailModal.tsx
