+++ app/[locale]/academics/helper/CourseList.tsx
