+++ app/[locale]/academics/helper/CourseListHeader.tsx
