+++ app/[locale]/academics/helper/CourseListRow.tsx
