+++ app/[locale]/academics/helper/CourseRow.tsx
