+++ app/[locale]/academics/helper/ScholarshipRow.tsx
