+++ app/[locale]/academics/undergraduate/courses/CourseToolbar.tsx
