+++ app/[locale]/academics/undergraduate/courses/page.tsx
@@ -1,7 +1,8 @@
 import { getCourses } from '@/apis/academics';
 
-import CoursePageContent from '@/components/academics/CoursePageContent';
-import RoadMapButton from '@/components/academics/RoadMapButton';
+import RoadMapButton from '@/app/[locale]/academics/helper/RoadMapButton';
+import CoursePageContent from '@/app/[locale]/academics/undergraduate/courses/CoursePageContent';
+
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
 import { Course } from '@/types/academics';
