+++ app/[locale]/academics/undergraduate/degree-requirements/page.tsx
@@ -1,6 +1,7 @@
 import { getDegreeRequirements } from '@/apis/academicsServer';
 
-import DegreeRequirementsBody from '@/components/academics/DegreeRequirements';
+import DegreeRequirementsBody from '@/app/[locale]/academics/helper/DegreeRequirements';
+
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
 export default async function UndergraduteDegreeRequirementsPage() {
