--- components/academics/CurriculumBody.tsx
@@ -1,27 +1,0 @@
-'use client';
-
-import { useState } from 'react';
-
-import { Curriculum } from '@/types/academics';
-
-import Dropdown from '../common/form/Dropdown';
-import HTMLViewer from '../editor/HTMLViewer';
-
-export default function CurriculumBody({ data }: { data: Curriculum[] }) {
-  const [selectedcurriculumIndex, setSelectedcurriculumIndex] = useState(0);
-  return (
-    <>
-      <div className="mt-8 flex flex-col">
-        <div className="flex flex-row items-center gap-2">
-          <p className=" text-sm font-normal">입학연도</p>
-          <Dropdown
-            contents={data.map((item) => item.year.toString())}
-            selectedIndex={selectedcurriculumIndex}
-            onClick={setSelectedcurriculumIndex}
-          />
-        </div>
-      </div>
-      <HTMLViewer htmlContent={data[selectedcurriculumIndex].description} className="mt-7" />
-    </>
-  );
-}
